etl/requirements.txt
backoff==2.2.1
elasticsearch==8.5.3
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pydantic==1.10.2
